requirements.txt
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
PyJWT
loguru==0.7.2
MarkupSafe==2.1.5
pycparser==2.22
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.4
